// src/modules/edit/types.ts
export interface EditParams {
  maxSize: number
  transparentBackground: boolean
  backgroundColor: string
  cornerRadius: number
  circleCrop: boolean
  strokeWidth: number
  strokeColor: string
  shadowEnabled: boolean
  shadowBlur: number
  shadowOffsetX: number
  shadowOffsetY: number
  shadowColor: string
}